### 2021-06-16 0a07b47
add random_subset method and tests
