+++ R/RcppExports.R
@@ -65,6 +65,10 @@ filter_bitset_bitset <- function(b, other) {
     .Call(`_individual_filter_bitset_bitset`, b, other)
 }
 
+bitset_random_subset <- function(b, N) {
+    invisible(.Call(`_individual_bitset_random_subset`, b, N))
+}
+
 create_categorical_variable <- function(categories, values) {
     .Call(`_individual_create_categorical_variable`, categories, values)
 }
