### 2021-06-17 49a1077
change random_subset -> choose and add suggested test
