### 2021-06-18 #3377 @mikldk
@giovannic:

- [ ] Can you please fix the missing DOI?
- [ ] Currently, the paper is too long (see the [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)) -- there are lots of good material. Have you considered moving some of the material to vignettes?
