### 2021-06-18 #3377 @slwu89
Hi @mikldk

I hope you don't mind me replying here; @giovannic has brilliantly led the software development but I was responsible for much of the paper text. I'll update those missing DOIs shortly.

I can shorten the paper by getting rid of all text beginning with the "Overview" on Page 4 and continuing to the end (although retaining Figure 1, which I think is quite important). I suspect though that even getting rid of 4 pages of content will still leave us above 1000 words; is a slight exceedance of that OK with the journal? I've found in the past some collaborators are reluctant to use some software if it's not accompanied by something that looks vaguely like a research article. I think keeping the sections (Summary, Statement of Need, Design Principles, and State of the Field + Figure 1) would be enough to get "buy-in" on using the software.

Also, do you have any suggestions how I can shorten the "State of the Field"? It consumes quite a bit of text because epidemic modelling has quite a few research software already (although I still think we have something new and useful to contribute).

Thanks!
