### 2021-06-20 #3377 @mikldk
@slwu89 Maybe it is worth considering converting it to a preprint on e.g. [arXiv](https://arxiv.org/)? And refer to that from this JOSS paper for details?
