### 2021-06-21 #3377 @giovannic
> any suggestions how I can shorten the "State of the Field"?

The most popular and relevant examples seem to be:

 * General
   * Repast
   * Mesa
   * NetLogo
   * R-based
     * RNetLogo
     * SpaDES
 * Epi-based
   * SimInf
   * EpiModel

For the others, we can either remove completely or cite. I can make a PR to cut this down
