### 2021-06-21 f749624
Reduced count in State of the Field
