+++ joss/paper.md
@@ -117,13 +117,11 @@ significantly enhancing the extensibility of `individual`'s API.
 
 # State of the field
 
-There are currently a variety of software libraries for epidemiological simulation,
-both in R and other programming languages. However, we are not yet aware
-of a similar package written especially for epidemiology which allows 
-simulation of generic infectious processes by providing users 
-with a set of specialized data types and methods without making assumptions about 
+There are many software libraries for epidemiological simulation,
+both in R and other programming languages. However, R lacks package written for
+the simulation of generic infectious processes, without making assumptions about 
 model structure (e.g.; network, metapopulation, lattice grid, etc). Additionally,
-none of the reviewed software used bitset objects to optimize certain operations.
+none of the reviewed software used bitsets for speed improvements.
 
 ## Non R Software
 
@@ -134,61 +132,49 @@ platform for specifying and simulation agent-based models.
 
 Various software libraries also exist specifically for epidemiological simulation.
 EpiFire [@Hladish:2012] and SimpactCyan [@Liesenborgs:2019] are C++ libraries
-for simulation of epidemics on networks. Numerus Model Builder [@Getz:2018] 
-and NOVA [@Salter:2013] use a domain specific language to define models. 
-EMOD [@Bershteyn:2018] is a highly successful and complex modeling 
-software in C++ which relies on JSON configuration files to interact with.
-Pathogen.jl [@Angevaare:2020] is a package for individual based simulation of common
-compartmental models.
+for simulation of epidemics on networks. Other software provide alternative
+interfaces for client models, including [@Getz:2018], [@Salter:2013],
+[@Bershteyn:2018], [@Angevaare:2020].
 
 ## R Software
 
 ### General R Packages
 
 Several R packages provide interfaces to other software libraries.
-The nlrx package provides an R interface to NetLogo [@Salecker:2019] to set up reproducible
-experiments and focuses on sensitivity analysis, and RNetLogo is a more basic interface [@Thiele:2014].
-NetLogoR [@Bauduin:2019] is a complete translation of NetLogo into R.
-For the Repast library, RRepast provides a sophisticated interface [@Garcia:2016].
-Finally, simecol [@Petzoldt:2007] provides classes and methods to enhance
-reproducibility of ecological models.
+The nlrx and NetLogoR packages provide an R interfaces to NetLogo [@Salecker:2019],
+[@Thiele:2014].
+Whereas NetLogoR [@Bauduin:2019] is a complete translation of NetLogo into R.
+RRepast provides a sophisticated interface [@Garcia:2016]. Finally, simecol
+[@Petzoldt:2007] provides classes and methods to enhance reproducibility of
+ecological models.
 
-Generic individual based simulation packages in R include the simmeR package for 
-discrete event simulation [@Ucar:2017] which also utilizes a `R6` interface to 
-lined C++ but is specialized for queueing type models.
-SpaDES [@Mcintire:2021] is a generic discrete event simulator designed for
-ecological models on raster landscapes. 
-IBMPopSim [@Giorgi:2020] is a sophisticated R package for simulation of 
-continuous time individual based models, but requires uses to input C++ code
+Generic individual based simulation packages in R include
+IBMPopSim [@Giorgi:2020], ibm [@Oliveros:2016] and ibmcraftr [@Tun:2016].
+IBMPopSim is the most sophisticated, but requires users to input C++ code
 as a string which is then compiled, making it difficult to interface with the
 existing R ecosystem.
 
-Other packages include ibm [@Oliveros:2016] which gives examples of individual based models
-in R but does not provide a generic interface, and ibmcraftr [@Tun:2016].
[…]
