### 2021-06-21 728931c
Limited SoF to R epi models
