### 2021-06-22 #3377 @slwu89
> @slwu89 I've tried to focus the "State of the Field" on R based, IBMs in Epi... [mrc-ide/individual#112](https://github.com/mrc-ide/individual/pull/112)

Thanks @giovannic! That's a great help.

@mikldk would something like the tidyverse JOSS paper be an appropriate length? https://joss.theoj.org/papers/10.21105/joss.01686 I'll go through and do some further editing after we merge the edits to the "State of the Field" section https://github.com/mrc-ide/individual/pull/112
