### 2021-06-22 a98e984
add missing DOIs
