+++ joss/paper.bib
@@ -4,7 +4,8 @@
   journal={Complex systems science in biomedicine},
   pages={33--114},
   year={2006},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-0-387-33532-2_2}
 }
 
 @article{Ajelli:2010,
@@ -111,7 +112,8 @@
   number={1},
   pages={1--12},
   year={2012},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/1471-2105-13-76}
 }
 
 @article{North:2013,
@@ -122,7 +124,8 @@
   number={1},
   pages={1--26},
   year={2013},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/2194-3206-1-3}
 }
 
 @inproceedings{Masad:2015,
@@ -131,7 +134,8 @@
   booktitle={14th PYTHON in Science Conference},
   pages={53--60},
   year={2015},
-  organization={Citeseer}
+  organization={Citeseer},
+  doi={10.25080/majora-7b98e3ed-009}
 }
 
 @techreport{Wilensky:1999,
@@ -154,10 +158,16 @@
 }
 
 @article{Ucar:2017,
-  title={simmer: Discrete-Event simulation for R},
-  author={Ucar, Inaki and Smeets, Bart and Azcorra, Arturo},
-  journal={arXiv preprint arXiv:1705.09746},
-  year={2017}
+   author = {Iñaki Ucar and Bart Smeets and Arturo Azcorra},
+   title = {simmer: Discrete-Event Simulation for R},
+   journal = {Journal of Statistical Software, Articles},
+   volume = {90},
+   number = {2},
+   year = {2019},
+   issn = {1548-7660},
+   pages = {1--30},
+   doi = {10.18637/jss.v090.i02},
+   url = {https://www.jstatsoft.org/v090/i02}
 }
 
 @software{Mcintire:2021,
@@ -184,7 +194,8 @@
   number={11},
   pages={1841--1849},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/ecog.04516}
 }
 
 @Article{Salecker:2019,
@@ -205,7 +216,8 @@
   volume={25},
   pages={9--19},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.epidem.2018.06.001}
 }
 
 @article{Salter:2013,
@@ -215,7 +227,8 @@
   volume={18},
   pages={1784--1793},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.procs.2013.05.347}
 }
 
 @article{Thiele:2014,
@@ -247,7 +260,8 @@
   number={1},
   pages={1--13},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1101/440834}
 }
 
 @article{Bauer:2016,
@@ -258,7 +272,8 @@
   number={4},
   pages={438--453},
   year={2016},
-  publisher={SAGE Publications Sage UK: London, England}
+  publisher={SAGE Publications Sage UK: London, England},
+  doi={10.1177/1094342016635723}
 }
 
 @article{Vahdati:2019,
@@ -349,7 +364,8 @@ Level Models for Epidemics},
   volume={22},
   number={9},
   pages={1--31},
-  year={2007}
+  year={2007},
+  doi={10.18637/jss.v022.i09}
 }
 
 @Manual{Handcock:statnet,
@@ -366,7 +382,8 @@ Level Models for Epidemics},
[…]
