### 2021-06-22 f45918b
significantly shorten paper
