+++ joss/paper.md
@@ -26,32 +26,11 @@ bibliography: paper.bib
 
 # Summary
 
-Complex stochastic models are a crucial tool for many tasks 
-in public health, and especially in infectious disease epidemiology [@Ganyani:2021]. 
-Such models can formalize theory, generate synthetic data, evaluate counterfactual 
-scenarios, forecast trends, and be used for statistical inference. Individual-based
-models (IBMs) are a way to design disaggregated simulation models, usually contrasted
-with mathematical or equation-based models, which may model a density or concentration
-of individuals, or otherwise lump individuals with similar attributes together in some
-way [@Shalizi:2006]. For modeling finite numbers of individuals with significant
-heterogeneity in characteristics, and dynamics which could depend on such 
-characteristics in complex ways, IBMs are a natural modeling
-choice whereas a representation using mathematical models (especially compartmental models) would be cumbersome
-or impossible [@Willem:2017]. Even if an aggregated representation were feasible, there are many 
-reasons why an individual-based representation is to be preferred. Synthetic data
-may need to produce individual level outcomes, which aggregated models by their very 
-nature are unable to provide [@Tracy:2018]. Other complexities, such as when events occur after
-a random delay whose distribution differs from a Markovian
-one, mean even aggregated models will need to store individual completion times,
-necessitating more complex simulation algorithms and data structures; in such
-cases it is often more straightforward to adopt an individual-based representation
-from the start.
-
 `individual` is an R package which provides users a set of useful primitive elements
-for specifying individual based simulation models, with special attention models
+for specifying individual based models (IBMs), with special attention to models
 for infectious disease epidemiology. 
-Users build models by specifying variables for each characteristic of individuals 
-in the simulated population by using data structures exposed by the package.
+Users build models by specifying variables for each variable describing individuals 
+in the simulated population using data structures exposed by the package.
 The package provides efficient methods for finding
 subsets of individuals based on these variables, or cohorts. Cohorts can then
 be targeted for variable updates or scheduled for events.
@@ -62,72 +41,52 @@ model of a system into executable code, which is fast and memory efficient.
 
 # Statement of need
 
-In many applications, but especially epidemiology, individual-based models 
+Complex stochastic models are a crucial tool for many tasks 
+in infectious disease epidemiology [@Ganyani:2021].
+Such models can formalize theory, generate synthetic data, evaluate counterfactual 
+scenarios, forecast trends, and be used for statistical inference. IBMs are a way to 
+design disaggregated simulation models, usually contrasted
+with mathematical, which may model a density or concentration
+of individuals, or otherwise lump individuals with similar attributes together in some
+way [@Shalizi:2006]. For modeling finite numbers of individuals with significant
+between-individual heterogeneity and complex dynamics, IBMs are a natural modeling
+choice when a representation using mathematical models would be cumbersome
+or impossible [@Willem:2017]. Even if an aggregated representation were feasible, there are many 
+reasons why an individual-based representation is to be preferred. Synthetic data
+may need to produce individual level outcomes, which aggregated models by their very 
+nature are unable to provide [@Tracy:2018]. Other complexities, such as when events occur after
+a random delay whose distribution differs from a Markovian
+one, mean even aggregated models will need to store individual completion times,
+necessitating more complex simulation algorithms and data structures; in such
+cases it is often more straightforward to adopt an individual-based representation
+from the start.
+
+For practical use, individual-based models 
 need to balance comprehensibility and speed. A fast model whose code is only
 understood by the author can be difficult to use as a basis for scientific
-exploration, which necessitates the development of various similar models to
[…]
