### 2021-06-22 5394480
fix typos, add a few more DOIs
