+++ joss/paper.bib
@@ -238,7 +238,8 @@
   volume={58},
   number={2},
   pages={1--41},
-  year={2014}
+  year={2014},
+  doi={10.18637/jss.v058.i02}
 }
 
 @article{Bershteyn:2018,
@@ -249,7 +250,8 @@
   number={5},
   pages={fty059},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/femspd/fty059}
 }
 
 @article{Liesenborgs:2019,
@@ -425,7 +427,8 @@ Level Models for Epidemics},
   number={8},
   pages={1002--1007},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/2041-210X.13422}
 }
 
 @article{Rizzi:2018,
