### 2021-06-23 #3377 @mikldk
@slwu89 The `tidyverse` paper is not the normal (...and has approx 1700 citations already). I think instead you should focus on our general guidelines: <https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain>
