+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   Contributing: Contributing.html
   Performance: Performance.html
   Tutorial: Tutorial.html
-last_built: 2021-06-22T08:07Z
+last_built: 2021-06-23T09:58Z
 
