### 2021-06-24 #3377 @slwu89
Thanks @mikldk, I've done a substantial edit here (https://github.com/mrc-ide/individual/pull/113) to reduce the length of the paper to something more acceptable, as well as adding all those missing DOIs.

[sample pdf](https://github.com/openjournals/joss-reviews/files/6712326/hgbphrqjg0ch4vvih2jz.pdf)
