### 2021-07-01 5384d83
add check if 'filter_bitset' gets empty value for 'other'
