### 2021-07-08 11c1353
get rid of CRAN check NOTEs
