### 2021-07-08 e6c9ab4
add ImportFrom to targeted_event.R for consistency
