### 2021-07-16 92a2d4b
CRAN test fixes:
 * fix signed compiler warnings on certain platforms
 * fix suspected UBSAN issue (cannot recreate on rocker)
