+++ R/double_variable.R
@@ -85,6 +85,7 @@ DoubleVariable <- R6Class(
           index <- index$to_vector()
         }
         if (length(index) != 0) {
+          stopifnot(all(index > 0))
           double_variable_queue_update(
             self$.variable,
             values,
