+++ inst/include/IterableBitset.h
@@ -479,7 +479,7 @@ inline void bitset_sample_internal(
         false // one based
     );
     std::sort(to_remove.begin(), to_remove.end());
-    auto bitset_i = 0u;
+    auto bitset_i = 0;
     auto bitset_it = b.cbegin();
     for (auto i : to_remove) {
       while(bitset_i != i) {
