### 2021-07-19 #3377 @SteRoe
Dear all,

yes, I'm interested in this review.

Best, Stefan

Am 18.07.2021 um 15:04 schrieb Mikkel Meyer Andersen:
>
> @seabbs <https://github.com/seabbs>, @strengejacke
> <https://github.com/strengejacke>: Would you be interested in
> reviewing this submission to The Journal of Open Source Software
> <https://joss.theoj.org/>? Reviews are open and based on a checklist.
> The reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If
> you have any questions or concerns please let me know.
>
