### 2021-07-19 592cde2
bitset choose now does nothing for k = n
