### 2021-07-26 #3539 @whedon
[REVIEW]: individual: An R package for individual-based epidemiological models
**Submitting author:** @giovannic (<a href="http://orcid.org/0000-0002-7024-1200">Giovanni Charles</a>)
**Repository:** <a href="https://github.com/mrc-ide/individual" target ="_blank">https://github.com/mrc-ide/individual</a>
**Version:** v0.1.6

- [x] #3591
