### 2021-07-26 #3539 @SteRoe
@mikldk: This is my first JOSS review. Filled out the check list, but I'm not sure how to submit it. Most notably I found no statement on examples and tests, so I left these checks open. Other points are fine. How to proceed?
