### 2021-07-26 #3539 @mikldk
@SteRoe Thanks for asking! The check items are not to be submitted, once they are checked that is saved and all is good. The ones you do not think are okay/not satisfied you can address here in this issue to the author (minor things), or preferably create issues in the submission's repository and mention this review issue so that the issues are linked.
