### 2021-07-26 #3539 @giovannic
@mikldk @SteRoe Thanks for bringing that up! Not sure how we declare these to JOSS but we have extensive automated tests ([here's our latest coverage](https://github.com/mrc-ide/individual/pull/127#issuecomment-885484195)), and our documentation has an example in the [tutorial](https://mrc-ide.github.io/individual/articles/Tutorial.html) (which had to be removed from the paper for brevity).
