### 2021-07-26 #3539 @giovannic
Some other potentially useful links...

[Installation instructions](https://github.com/mrc-ide/individual#installation)
[Contributing guidelines](https://mrc-ide.github.io/individual/articles/Contributing.html)

;)
