### 2021-07-27 ba09b93
Add code coverage and tutorial links to joss paper
