### 2021-07-27 #3539 @giovannic
I've [added](https://github.com/mrc-ide/individual/commit/ba09b93f3458d9f2ec629d08ea89ba210b1bde7d) a short note to the Licensing and Availability section.
