### 2021-07-29 b302a5a
Add logo to README
