+++ README.md
@@ -1,8 +1,11 @@
-<!-- badges: start -->
-[![R build status](https://github.com/mrc-ide/individaul/workflows/R-CMD-check/badge.svg)](https://github.com/mrc-ide/individual/actions)
-<!-- badges: end -->
+# Individual <img src='man/figures/logo.png' align="right" height="139" />
 
-# Individual
+<!-- badges: start -->
+[![R build status](https://github.com/mrc-ide/individual/workflows/R-CMD-check/badge.svg)](https://github.com/mrc-ide/individual/actions)
+[![codecov.io](https://codecov.io/github/mrc-ide/individual/coverage.svg)](https://codecov.io/github/mrc-ide/individual)
+[![CRAN](https://www.r-pkg.org/badges/version/individual)](https://cran.r-project.org/package=individual)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+<!-- badges: end -->
 
 An R package for specifying and simulating individual based models.
 
@@ -56,5 +59,49 @@ learn more about how to speed up your model in `vignette("Performance")`.
 Thank you! Please refer to the vignette on `vignette("Contributing")` for info on how to
 contribute :)
 
-## License
-[MIT](https://choosealicense.com/licenses/mit/)
+## Alternatives
+
+### Non R Software
+
+ - Repast
+ - Mesa
+ - NetLogo
+ - Agents.jl
+
+### Non R Software for Epi
+
+ - EpiFire
+ - SimpactCyan
+ - Numerus Model Builder
+ - NOVA
+ - EMOD
+ - Pathogen.jl, a package for individual based simulation of common compartmental models.
+
+### General R Packages
+
+ - nlrx, RNetLogo, NetLogoR are NetLogo interfaces
+ - RRepast is a repast interface
+ - simecol, provides classes and methods to enhance reproducibility of ecological models.
+
+### R based DES
+
+ - simmeR
+ - SpaDES
+
+### R based IBMs
+
+ - IBMPopSim
+ - ibm
+ - ibmcraftr
+
+### R based Epi
+
+ - EpiModel 
+ - SimInf
+ - hybridModels
+ - epinet
+ - EpiDynamics
+ - nosoi, generate synthetic data for phylogenetic analysis
+ - EpiILMCT
+ - EpiILM
+ - SPARSEMODr
