### 2021-08-06 3de7209
test for IntegerVariable queue update with bitset passed to c++
