+++ R/integer_variable.R
@@ -134,6 +134,64 @@ IntegerVariable <- R6Class(
         }
       }
     },
+    
+    #' @description Queue an update for a variable. There are 4 types of variable update:
+    #'
+    #' \enumerate{
+    #'  \item{Subset update: }{The argument \code{index} represents a subset of the variable to
+    #' update. The argument \code{values} should be a vector whose length matches the size of \code{index},
+    #' which represents the new values for that subset.}
+    #'  \item{Subset fill: }{The argument \code{index} represents a subset of the variable to
+    #' update. The argument \code{values} should be a single number, which fills the specified subset.}
+    #'  \item{Variable reset: }{The index vector is set to \code{NULL} and the argument \code{values}
+    #' replaces all of the current values in the simulation. \code{values} should be a vector
+    #' whose length should match the size of the population, which fills all the variable values in
+    #' the population}
+    #'  \item{Variable fill: }{The index vector is set to \code{NULL} and the argument \code{values}
+    #' should be a single number, which fills all of the variable values in 
+    #' the population.}
+    #' }
+    #' @param values a vector or scalar of values to assign at the index
+    #' @param index is the index at which to apply the change, use \code{NULL} for the
+    #' fill options. If using indices, this may be either a vector of integers or
+    #' a \code{\link[individual]{Bitset}}.
+    queue_update_bitset = function(values, index = NULL) {
+      stopifnot(is.numeric(values))
+      if(is.null(index)){
+        if(length(values) == 1){
+          integer_variable_queue_fill(
+            self$.variable,
+            values
+          )
+        } else {
+          integer_variable_queue_update(
+            self$.variable,
+            values,
+            numeric(0)
+          )
+        }
+      } else {
+        if (inherits(index, 'Bitset')) {
+          if (index$size() > 0) {
+            integer_variable_queue_update_bitset(
+              self$.variable,
+              values,
+              index$.bitset
+            )
+          }
+        } else {
+          if (length(index) > 0) {
+            stopifnot(all(index > 0))
+            integer_variable_queue_update(
+              self$.variable,
+              values,
+              index
+            )
+          }          
+        }
+
+      }
+    },
 
     .update = function() integer_variable_update(self$.variable)
   )
