+++ src/RcppExports.cpp
@@ -9,6 +9,11 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // create_bitset
 Rcpp::XPtr<individual_index_t> create_bitset(size_t size);
 RcppExport SEXP _individual_create_bitset(SEXP sizeSEXP) {
@@ -160,6 +165,17 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const Rcpp::XPtr<individual_index_t> >::type b(bSEXP);
     rcpp_result_gen = Rcpp::wrap(bitset_to_vector(b));
+    return rcpp_result_gen;
+END_RCPP
+}
+// bitset_to_vector_call_internal
+std::vector<size_t> bitset_to_vector_call_internal(const Rcpp::XPtr<individual_index_t> b);
+RcppExport SEXP _individual_bitset_to_vector_call_internal(SEXP bSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Rcpp::XPtr<individual_index_t> >::type b(bSEXP);
+    rcpp_result_gen = Rcpp::wrap(bitset_to_vector_call_internal(b));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -639,6 +655,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// integer_variable_get_values_at_index_vector_no_convert
+std::vector<int> integer_variable_get_values_at_index_vector_no_convert(Rcpp::XPtr<IntegerVariable> variable, std::vector<size_t> index);
+RcppExport SEXP _individual_integer_variable_get_values_at_index_vector_no_convert(SEXP variableSEXP, SEXP indexSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::XPtr<IntegerVariable> >::type variable(variableSEXP);
+    Rcpp::traits::input_parameter< std::vector<size_t> >::type index(indexSEXP);
+    rcpp_result_gen = Rcpp::wrap(integer_variable_get_values_at_index_vector_no_convert(variable, index));
+    return rcpp_result_gen;
+END_RCPP
+}
 // integer_variable_get_index_of_set_vector
 Rcpp::XPtr<individual_index_t> integer_variable_get_index_of_set_vector(Rcpp::XPtr<IntegerVariable> variable, std::vector<int> values_set);
 RcppExport SEXP _individual_integer_variable_get_index_of_set_vector(SEXP variableSEXP, SEXP values_setSEXP) {
@@ -733,6 +761,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const std::vector<int> >::type value(valueSEXP);
     Rcpp::traits::input_parameter< std::vector<size_t> >::type index(indexSEXP);
     integer_variable_queue_update(variable, value, index);
+    return R_NilValue;
+END_RCPP
+}
+// integer_variable_queue_update_bitset
[…]
