### 2021-08-06 374a78c
out of struct definitions of methods for IntegerVariable
