### 2021-08-06 65c0a78
pass index vector by reference to IntegerVariable methods
