### 2021-08-06 ea8f9b0
virtual dtor for IntegerVariable
