### 2021-08-08 9b48a36
DoubleVariable do not return reference to values
