### 2021-08-08 3b3b9ff
catch bug with out of range index passed to get_value; add more tests
