+++ inst/include/DoubleVariable.h
@@ -38,87 +38,12 @@ struct DoubleVariable : public Variable {
     virtual std::vector<double> get_values(const individual_index_t& index) const;
     virtual std::vector<double> get_values(const std::vector<size_t>& index) const;
 
-    // get indices of individuals whose value is in some [a,b]
-    virtual individual_index_t get_index_of_range(
-        const double a, const double b
-    ) const {
-        
-        auto result = individual_index_t(size);
-        for (auto i = 0u; i < values.size(); ++i) {
-            if( !(values[i] < a) && !(b < values[i]) ) {
-                result.insert(i);
-            }
-        }
-        
-        return result;
+    virtual individual_index_t get_index_of_range(const double a, const double b) const;
+    virtual size_t get_size_of_range(const double a, const double b) const;
 
-    }
-
-    // get indices of individuals whose value is in some [a,b]
-    virtual size_t get_size_of_range(
-        const double a, const double b
-    ) const {
-        
-        size_t result = std::count_if(values.begin(), values.end(), [&](const double v) -> bool {
-            return !(v < a) && !(b < v);
-        });
-
-        return result;
-
-    }
-
-    virtual void queue_update(
-        const std::vector<double>& values,
-        const std::vector<size_t>& index
-    ) {
-        if (values.size() > 1 && values.size() < size && values.size() != index.size()) {
-            Rcpp::stop("Mismatch between value and index length");
-        }
-        for (auto i : index) {
-            if (i < 0 || i >= size) {
-                Rcpp::stop("Index out of bounds");
-            }
-        }
-        updates.push({ values, index });
-    }
-
-    virtual void update() override {
-        while(updates.size() > 0) {
-            const auto& update = updates.front();
-            const auto& values = update.first;
-            const auto& index = update.second;
-            if (values.size() == 0) {
-                return;
-            }
-
-            auto vector_replacement = (index.size() == 0);
-            auto value_fill = (values.size() == 1);
-
-            auto& to_update = this->values;
-
-            if (vector_replacement) {
-                // For a full vector replacement
-                if (value_fill) {
-                    to_update = std::vector<double>(size, values[0]);
-                } else {
-                    to_update = values;
-                }
-            } else {
-                if (value_fill) {
-                    // For a fill update
-                    for (auto i : index) {
-                        to_update[i] = values[0];
-                    }
-                } else {
-                    // Subset assignment
-                    for (auto i = 0u; i < index.size(); ++i) {
-                        to_update[index[i]] = values[i];
-                    }
-                }
-            }
-            updates.pop();
-        }
-    }
+    virtual void queue_update(const std::vector<double>& values, const std::vector<size_t>& index);
+    virtual void update() override;
+    
 };
 
 
@@ -146,13 +71,94 @@ inline std::vector<double> DoubleVariable::get_values(const individual_index_t& 
 }
 
 inline std::vector<double> DoubleVariable::get_values(const std::vector<size_t>& index) const {
+    
     auto result = std::vector<double>(index.size());
-    auto result_i = 0u;
-    for (auto i : index) {
-        result[result_i] = values.at(i);
-        ++result_i;
+    for (auto i = 0u; i < index.size(); ++i) {
+        result[i] = values.at(index[i]);
     }
     return result;
+}
+
+// get indices of individuals whose value is in some [a,b]
+inline individual_index_t DoubleVariable::get_index_of_range(
+        const double a, const double b
+) const {
+    
+    auto result = individual_index_t(size);
+    for (auto i = 0u; i < values.size(); ++i) {
+        if( !(values[i] < a) && !(b < values[i]) ) {
+            result.insert(i);
+        }
+    }
+    
[…]
