+++ src/integer_variable.cpp
@@ -36,17 +36,6 @@ std::vector<int> integer_variable_get_values_at_index(
 
 //[[Rcpp::export]]
 std::vector<int> integer_variable_get_values_at_index_vector(
-    Rcpp::XPtr<IntegerVariable> variable,
-    std::vector<size_t> index
-    ) {
-    decrement(index);
-    auto bitmap = individual_index_t(variable->size);
-    bitmap.insert_safe(index.begin(), index.end());
-    return variable->get_values(bitmap);
-}
-
-//[[Rcpp::export]]
-std::vector<int> integer_variable_get_values_at_index_vector_no_convert(
         Rcpp::XPtr<IntegerVariable> variable,
         std::vector<size_t> index
 ) {
