+++ R/double_variable.R
@@ -82,15 +82,22 @@ DoubleVariable <- R6Class(
         }
       } else {
         if (inherits(index, 'Bitset')) {
-          index <- index$to_vector()
-        }
-        if (length(index) != 0) {
-          stopifnot(all(index > 0))
-          double_variable_queue_update(
-            self$.variable,
-            values,
-            index
-          )
+          if (index$size() > 0){
+            double_variable_queue_update_bitset(
+              self$.variable,
+              values,
+              index$.bitset
+            )
+          }
+        } else {
+          if (length(index) != 0) {
+            stopifnot(all(index > 0))
+            double_variable_queue_update(
+              self$.variable,
+              values,
+              index
+            )
+          }
         }
       }
     },

[…]
