### 2021-08-08 7e7ee55
split variable test by class
