### 2021-08-09 b97d820
remove comparison method from Bitset
