+++ R/RcppExports.R
@@ -57,10 +57,6 @@ bitset_to_vector <- function(b) {
     .Call(`_individual_bitset_to_vector`, b)
 }
 
-bitset_to_vector_call_internal <- function(b) {
-    .Call(`_individual_bitset_to_vector_call_internal`, b)
-}
-
 filter_bitset_vector <- function(b, other) {
     .Call(`_individual_filter_bitset_vector`, b, other)
 }
