### 2021-08-09 b48ade1
clean up integer and double variable header files; remove sstream from includes
