### 2021-08-09 fba6441
categorical variable get size methods need to return size_t
