### 2021-08-09 c01ffce
simpler get index of for categorical variable with one category
