### 2021-08-09 166d068
remove bitset include from IterableBitset.h
