+++ inst/include/IntegerVariable.h
@@ -31,7 +31,7 @@ struct IntegerVariable : public Variable {
     std::vector<int> values;
 
     IntegerVariable(const std::vector<int>& values);
-    virtual ~IntegerVariable();
+    virtual ~IntegerVariable() = default;
 
     virtual std::vector<int> get_values() const;
     virtual std::vector<int> get_values(const individual_index_t& index) const;
@@ -55,7 +55,7 @@ inline IntegerVariable::IntegerVariable(const std::vector<int>& values)
     : size(values.size()), values(values)
 {}
 
-inline IntegerVariable::~IntegerVariable() {};
+// inline IntegerVariable::~IntegerVariable() {};
 
 //' @title get all values
 inline std::vector<int> IntegerVariable::get_values() const {
