### 2021-08-09 8205290
don't need to check size_t < 0
