+++ inst/include/IntegerVariable.h
@@ -174,7 +174,7 @@ inline void IntegerVariable::queue_update(
         Rcpp::stop("Mismatch between value and index length");
     }
     for (auto i : index) {
-        if (i < 0 || i >= size) {
+        if (i >= size) {
             Rcpp::stop("Index out of bounds");
         }
     }
