### 2021-08-09 995186c
unneeded check for size_t < 0 in iterable bitset
