### 2021-08-09 mrc-ide/individual#132 @strengejacke
[JOSS] plain LICENSE file not showing OSI conform text
This is related to https://github.com/openjournals/joss-reviews/issues/3539

The plain text file for the LICENSE should contain this text:
https://opensource.org/licenses/MIT
