### 2021-08-09 mrc-ide/individual#133 @strengejacke
[JOSS] Statement of need
This issue is related to https://github.com/openjournals/joss-reviews/issues/3539.

A statement of need is missing in the software documentation (readme).
