### 2021-08-09 cbf951a
checked for undef behavior and valgrind; removed reshape2 from DESCRIPTION
