+++ R/integer_variable.R
@@ -27,6 +27,7 @@ IntegerVariable <- R6Class(
         return(integer_variable_get_values(self$.variable))
       }
       if (is.numeric(index)) {
+        stopifnot(all(index > 0))
         return(integer_variable_get_values_at_index_vector(self$.variable, index))
       }
       integer_variable_get_values_at_index(self$.variable, index$.bitset)
