### 2021-08-09 7692c11
remove commented out code
