+++ inst/include/IntegerVariable.h
@@ -179,7 +179,7 @@ inline void IntegerVariable::queue_update(
     updates.push({ values, index });
 }
 
-//' @title apply all queued state updates in LIFO order
+//' @title apply all queued state updates in FIFO order
 inline void IntegerVariable::update() {
     while(updates.size() > 0) {
         const auto& update = updates.front();
