### 2021-08-09 4b85553
add more tests for edge cases; check bitset size when updating
