### 2021-08-10 mrc-ide/individual#133 @giovannic
Hello,

Our package is now listed on CRAN which means it must conform to R package conventions. And their best practice is to treat the LICENSE file as a template:

> A name of a license “template” and a file containing specific variables. The most common case is MIT + file LICENSE, where the LICENSE file needs to contain two fields: the year and copyright holder.

- from [their docs](https://r-pkgs.org/license.html)

What do you suggest we do? The full text _does_ appear in LICENSE.md
