### 2021-08-10 mrc-ide/individual#132 @strengejacke
Ok, if this is a CRAN requirement and the md-file lists the full text, I think this ok then.
