### 2021-08-10 f0e5c44
add NEWS.md and update pkgdown reference organization
