### 2021-08-12 089312f
add statement of need to README.md
