### 2021-08-14 56e394a
add check to TargetedEvent for index > 0 otherwise will get undefined behavior warning on CRAN
