### 2021-08-14 e63d39c
check bitset of compatible size when scheduling TargetedEvent
