### 2021-08-14 bbb91d3
TargetedEvent schedule method for bitset and delay vector
