### 2021-08-14 1116cc8
add checks for bad delay input
