### 2021-08-14 322f374
overload round function for doubles with check for positive input
