### 2021-08-14 6bce6d8
use insert_safe for bitset in event.cpp for more comprehensible errors
