+++ inst/include/Event.h
@@ -1,12 +1,12 @@
 /*
- * Scheduler.h -> Event.h
+ *  Event.h
  *
  *  Created on: 19 May 2020
  *      Author: gc1610
  */
 
-#ifndef INST_INCLUDE_SCHEDULER_H_
-#define INST_INCLUDE_SCHEDULER_H_
+#ifndef INST_INCLUDE_EVENT_H_
+#define INST_INCLUDE_EVENT_H_
 
 #include "common_types.h"
 #include <Rcpp.h>
@@ -17,6 +17,8 @@
 using listener_t = std::function<void (size_t)>;
 using targeted_listener_t = std::function<void (size_t, const individual_index_t&)>;
 
+
+//' @title round a double, will error if input is negative or not finite
 inline size_t round_double(double x) {
     if (x < 0.0 || !std::isfinite(x)) {
         Rcpp::stop("delay must be >= 0");
@@ -25,6 +27,7 @@ inline size_t round_double(double x) {
     }
 }
 
+//' @title round a vector of doubles
 inline std::vector<size_t> round_delay(const std::vector<double>& delay) {
     auto rounded = std::vector<size_t>(delay.size());
     for (auto i = 0u; i < delay.size(); ++i) {
@@ -33,169 +36,251 @@ inline std::vector<size_t> round_delay(const std::vector<double>& delay) {
     return rounded;
 }
 
+//' @title abstract base class for events
 struct EventBase {
     size_t t = 1;
 
-    virtual void tick() {
-        ++t;
-    }
-
+    virtual void tick();
+    
     virtual bool should_trigger() = 0;
     virtual ~EventBase() = default;
 };
 
+//' @title increase time step by one
+inline void EventBase::tick() {
+    ++t;
+}
+
+
+//' @title a general event in the simulation
+//' @description This class provides functionality for general events which are 
+//' applied to all individuals in the simulation. It inherits from EventBase.
+//' It contains the following data members:
+//'     * t: current simulation time step
+//'     * simple_schedule: a set of times the event will fire
 struct Event : public EventBase {
 
     std::set<size_t> simple_schedule;
+    
+    virtual ~Event() = default;
 
-    virtual void process(Rcpp::XPtr<listener_t> listener) {
-        (*listener)(t);
-    }
+    virtual void process(Rcpp::XPtr<listener_t> listener);
+    virtual bool should_trigger() override;
+    virtual void tick() override;
 
-    virtual bool should_trigger() override {
-        return *simple_schedule.begin() == t;
-    }
-
-    virtual void tick() override {
-        simple_schedule.erase(t);
-        EventBase::tick();
-    }
-
-    virtual void schedule(std::vector<double> delays) {
-        for (auto delay : round_delay(delays)) {
-            simple_schedule.insert(t + delay);
-        }
-    }
-
-    virtual void clear_schedule() {
-        simple_schedule.clear();
-    }
-
-    virtual ~Event() {};
+    virtual void schedule(std::vector<double> delays);
+    virtual void clear_schedule();
+    
 };
 
+//' @title process an event by calling a listener
+inline void Event::process(Rcpp::XPtr<listener_t> listener) {
+    (*listener)(t);
+}
+
+//' @title should first event fire on this timestep?
+inline bool Event::should_trigger() {
+    return *simple_schedule.begin() == t;
+}
+
+//' @title delete current time step from simple_schedule and increase time step
+inline void Event::tick() {
+    simple_schedule.erase(t);
+    EventBase::tick();
+}
+
+//' @title schedule a vector of events
+inline void Event::schedule(std::vector<double> delays) {
+    for (auto delay : round_delay(delays)) {
+        simple_schedule.insert(t + delay);
+    }
+}
+
[…]
