### 2021-08-15 9a5476c
continue adding tests for targeted event
