### 2021-08-15 mrc-ide/individual#133 @giovannic
Resolved by #137
