+++ tests/testthat/test-events.R
@@ -59,327 +59,3 @@ test_that("events can be scheduled for the future", {
   mockery::expect_args(listener, 1, t = 3)
   mockery::expect_args(listener, 2, t = 4)
 })
-
-test_that("targeted events can be scheduled for the future (vector)", {
-  event <- TargetedEvent$new(10)
-  listener <- mockery::mock()
-  event$add_listener(listener)
-  event$schedule(c(2, 4), 2)
-
-  #time = 1
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-
-  #time = 2
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-
-  #time = 3
-  event$.process()
-  mockery::expect_called(listener, 1)
-  event$.tick()
-
-  #time = 4
-  event$.process()
-  mockery::expect_called(listener, 1)
-  expect_targeted_listener(listener, 1, t = 3, target = c(2, 4))
-})
-
-test_that("targeted events can be scheduled for the future (bitset)", {
-  event <- TargetedEvent$new(10)
-  listener <- mockery::mock()
-  event$add_listener(listener)
-  event$schedule(Bitset$new(10)$insert(c(2,4)), 2)
-  
-  #time = 1
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-  
-  #time = 2
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-  
-  #time = 3
-  event$.process()
-  mockery::expect_called(listener, 1)
-  event$.tick()
-  
-  #time = 4
-  event$.process()
-  mockery::expect_called(listener, 1)
-  expect_targeted_listener(listener, 1, t = 3, target = c(2, 4))
-})
-
-test_that("events can be scheduled for for a Real time (vector)", {
-  event <- TargetedEvent$new(10)
-  listener <- mockery::mock()
-  event$add_listener(listener)
-  event$schedule(c(2, 4), 1.9)
-
-  #time = 1
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-
-  #time = 2
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-
-  #time = 3
-  event$.process()
-  mockery::expect_called(listener, 1)
-  event$.tick()
-
-  #time = 4
-  event$.process()
-  mockery::expect_called(listener, 1)
-  expect_targeted_listener(listener, 1, t = 3, target = c(2, 4))
-})
-
-("events can be scheduled for for a Real time (bitset)", {
-  event <- TargetedEvent$new(10)
-  listener <- mockery::mock()
-  event$add_listener(listener)
-  event$schedule(Bitset$new(10)$insert(c(2,4)), 1.9)
-  
-  #time = 1
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-  
-  #time = 2
-  event$.process()
-  mockery::expect_called(listener, 0)
-  event$.tick()
-  
-  #time = 3
-  event$.process()
-  mockery::expect_called(listener, 1)
-  event$.tick()
-  
-  #time = 4
-  event$.process()
-  mockery::expect_called(listener, 1)
-  expect_targeted_listener(listener, 1, t = 3, target = c(2, 4))
-})
-
-test_that("you can schedule different times for a target population (vector)", {
-  event <- TargetedEvent$new(10)
-  listener <- mockery::mock()
-  event$add_listener(listener)
-  event$schedule(c(1, 2, 4, 8, 3), c(1, 3, 1, 2, 2))
-
-  #time = 1
-  event$.process()
[…]
