### 2021-08-15 a853a0a
figure out where event fails for test-targetedevent.R:463 to write a better error message and avoid undef behavior
