### 2021-08-15 7eb6c94
use insert_safe when inserting into bitset from vector to get more comprehensible error messages for bad input
