### 2021-08-15 623907e
behavior is to only schedule if target has more than 0 elements
