### 2021-08-15 9799ebf
add docs to insert_safe for iterable bitsets
