### 2021-08-15 ca2a08f
add complete tests for all inputs to TargetedEvent
