### 2021-08-16 281899c
include better checks for any time an R vector will be cast to vector<size_t> to avoid undefined behavior
