### 2021-08-16 f1a945d
update changelog (NEWS.md) and bump version
