### 2021-08-19 396c7ec
add test for numeric times for Event and clearer returns in integer/double variable
