+++ tests/testthat/test-events.R
@@ -59,3 +59,31 @@ test_that("events can be scheduled for the future", {
   mockery::expect_args(listener, 1, t = 3)
   mockery::expect_args(listener, 2, t = 4)
 })
+
+test_that("events can be scheduled for a real time", {
+  event <- Event$new()
+  listener <- mockery::mock()
+  event$add_listener(listener)
+  event$schedule(c(2.1, 3.1))
+  
+  #time = 1
+  event$.process()
+  mockery::expect_called(listener, 0)
+  event$.tick()
+  
+  #time = 2
+  event$.process()
+  mockery::expect_called(listener, 0)
+  event$.tick()
+  
+  #time = 3
+  event$.process()
+  mockery::expect_called(listener, 1)
+  event$.tick()
+  
+  #time = 4
+  event$.process()
+  mockery::expect_called(listener, 2)
+  mockery::expect_args(listener, 1, t = 3)
+  mockery::expect_args(listener, 2, t = 4)
+})
