### 2021-08-20 85fb817
Update bitset$not:
* Add inplace option
* Add deprecation warning
* Rename operator ~ to operator !
