### 2021-08-20 42321c5
update Tutorial for new  method
