+++ vignettes/Tutorial.Rmd
@@ -93,7 +93,7 @@ We note at this point would be possible to queue the recovery event at the same 
 recovery_process <- function(t){
   I <- health$get_index_of("I")
   already_scheduled <- recovery_event$get_scheduled()
-  I$and(already_scheduled$not())
+  I$and(already_scheduled$not(inplace = TRUE))
   rec_times <- rgeom(n = I$size(),prob = pexp(q = gamma * dt)) + 1
   recovery_event$schedule(target = I,delay = rec_times)
 }
