+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   Contributing: Contributing.html
   Performance: Performance.html
   Tutorial: Tutorial.html
-last_built: 2021-08-10T08:57Z
+last_built: 2021-08-20T16:13Z
 
