### 2021-09-24 3f26dfc
final revisions to the JOSS paper
