### 2021-09-27 #3539 @giovannic
Thanks @mikldk ,

We've published v1.6 [here](https://zenodo.org/record/5529932), with DOI: 10.5281/zenodo.5529932

Will do a final check of the paper...
