### 2021-10-04 #3539 @mikldk
@slwu89  @giovannic: At Zenodo you write `v1.6` (right hand sidebar), but the release at Github is `v0.1.6` (from Aug 20?)? Also, the names on the paper and at Zenodo do not match, please make consistent.
