### 2021-10-05 #3539 @mikldk
@giovannic:

1. In the Summary section, you write "individual based models (IBMs)" - would it be appropriate to also include the term "agent based models (ABM)"?
2. Sometimes you write "individual based" and others "individual-based". Please use the same form.
3. "General R Packages" vs "Design Principles" vs "State of the field" - please either use Title Case consistently, or don't use it :smile:.
4. Acknowledgements: Please end the section with a period. Also, please write what you acknowledge them for.
