### 2021-10-05 864f61e
Small JOSS fixes
