### 2021-10-05 8e4ea41
work out grid for bitset benchmark
