### 2021-10-05 #3539 @mikldk
@giovannic The title still says "[...] for individual based epidemiological models" - without hyphen?
