### 2021-10-05 ec4f1be
hyphen in JOSS paper title
