### 2021-10-05 #3539 @slwu89
@mikldk I've addressed that here https://github.com/mrc-ide/individual/commit/ec4f1be35e1e0e2a599db67c2a9120aea09ab630
