### 2021-10-05 #3539 @mikldk
@openjournals/joss-eics : Please note that the title has changed (a hyphen added to "individual-based"). I changed the issue title here, but I am not sure if it needs to be changed elsewhere, too.
