### 2021-10-05 183152b
work on random data gen for testing variables
