### 2021-10-05 2f36b01
integer variable for scalar value and bitset index
